requirements.txt
pillow
opencv-python
pysimplegui